server/requirements.txt
psycopg2-binary==2.9.9
flask==3.0.3
